requirements.txt
pycryptodomex
gmssl
requests